mcp/requirements.txt
fastmcp>=0.1.0
starlette>=0.27.0
uvicorn>=0.24.0
openai>=1.0.0
jinja2>=3.1.0
requests>=2.31.0 
databricks-sdk
markdown>=3.4.0